# docs/swagger.yaml
basePath: /
definitions:
...
  editor.Content_tags:
    properties:
      cover:
        items:
          type: string
        type: array
      desc:
        type: string
      flag:
        items:
          type: string
        type: array
      has_changed:
        type: boolean
      has_del:
        type: boolean
      key:
        type: string
      list_order:
        type: integer
      name:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  editor.DiscoverChannel:
    properties:
      icon:
        type: string
      isDel:
        type: boolean
      isShow:
        type: boolean
      listOrder:
        type: integer
      showPos:
        type: string
      title:
        additionalProperties:
          type: string
        type: object
      uuid:
        type: string
    type: object
...
  editorapi.CoursePriDetail:
    properties:
      content:
        properties:
          lesson:
            type: string
          sentence:
            type: string
          word:
            type: string
        type: object
      course_code:
        type: string
      desc:
        type: string
      features:
        items:
          properties:
            desc:
              type: string
            img:
              type: string
            list_order:
              type: integer
            title:
              type: string
          type: object
        type: array
      flag:
        type: string
      has_dict:
        type: boolean
      sound_actors:
        items:
          $ref: '#/definitions/editorapi.soundActor'
        type: array
      title:
        type: string
    type: object
  editorapi.FeildInfo:
    properties:
      data_from:
        type: string
      desc:
        type: string
      feild:
        type: string
      list_order:
        type: integer
      name:
        type: string
      sub_feilds:
        items:
          $ref: '#/definitions/editorapi.subFeildInfo'
        type: array
      type:
        type: string
    type: object
...
  editorapi.SearchWord:
    properties:
      listOrder:
        type: integer
      uuid:
        type: string
      word:
        type: string
    type: object
...
  editorapi.catalogEditInfo:
    properties:
      attr_tag:
        type: string
      cover:
        items:
          type: string
        type: array
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      goalTitle:
        type: string
      has_changed:
        type: boolean
      is_show:
        type: boolean
      list_order:
        type: integer
      name:
        type: string
      onlineState:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        additionalProperties:
          type: string
        type: object
      update_time:
        type: integer
    type: object
  editorapi.catalogEditParam:
    properties:
      catalog_info:
        $ref: '#/definitions/editorapi.catalogEditInfo'
        type: object
      uuid:
        type: string
    type: object
...
  editorapi.classWeixinEditPara:
    properties:
      courseCode:
        type: string
      id:
        type: string
      isShow:
        type: boolean
      weixinCode:
        type: string
      weixinNo:
        type: string
    type: object
...
  editorapi.contenType:
    properties:
      desc:
        type: string
      has_del:
        type: boolean
      list_order:
        type: integer
      model_key:
        type: string
      model_keys:
        items:
          type: string
        type: array
      name:
        type: string
      type:
        type: string
    type: object
...
  editorapi.contentEditParam:
    properties:
      content_model:
        description: 内容模型名
        type: string
      contents:
        description: 多个元素的文件内容
        type: string
      parent_uuid:
        description: 文件目录的UUID
        type: string
    type: object
  editorapi.contentEditParams:
    properties:
      content_info:
        $ref: '#/definitions/editorapi.contentInfo'
        type: object
      uuid:
        type: string
    type: object
...
  editorapi.contentInfo:
    properties:
      cover:
        items:
          type: string
        type: array
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      has_changed:
        type: boolean
      is_show:
        type: boolean
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        additionalProperties:
          type: string
        type: object
      update_time:
        type: integer
      version:
        type: string
    type: object
...
  editorapi.courseEditInfo:
    properties:
      course_type:
        type: integer
      cover:
        items:
          type: string
        type: array
      default_actor:
        type: string
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      has_dict:
        type: boolean
      is_show:
        type: boolean
      name:
        type: string
      sound_actors:
        items:
          $ref: '#/definitions/editorapi.soundActor'
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        additionalProperties:
          type: string
        type: object
    type: object
  editorapi.courseEditParam:
    properties:
      editInfo:
        $ref: '#/definitions/editorapi.courseEditInfo'
        type: object
      uuid:
        type: string
    type: object
...
  editorapi.imageAddMoreParam:
    properties:
      image_urls:
        items:
          type: string
        type: array
      names:
        items:
          type: string
        type: array
      tagKeys:
        items:
          type: string
        type: array
    type: object
  editorapi.imageAddParam:
    properties:
      image_url:
        type: string
      tagKeys:
        items:
          type: string
        type: array
    type: object
  editorapi.imageDelParam:
    properties:
      image_id:
        type: string
    type: object
  editorapi.imageEditParam:
    properties:
      desc:
        type: string
      image_id:
        type: string
      image_url:
        type: string
      tagKeys:
        items:
          type: string
        type: array
    type: object
  editorapi.imageSearchParam:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      searchType:
        description: 0模糊搜索，1精确搜索
        type: integer
      tagKey:
        type: string
      words:
        type: string
    type: object
  editorapi.imageTagParam:
    properties:
      tagKey:
        type: string
    type: object
...
  editorapi.langEdit:
    properties:
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      is_hot:
        type: boolean
      is_show:
        type: boolean
      list_order:
        type: integer
      title:
        additionalProperties:
          type: string
        type: object
      word_direction:
        type: string
    type: object
  editorapi.langEditPara:
    properties:
      lan_code:
        type: string
      lang_info:
        $ref: '#/definitions/editorapi.langEdit'
        type: object
    type: object
...
  editorapi.modelEdit:
    properties:
      desc:
        type: string
      feilds:
        items:
          $ref: '#/definitions/editorapi.FeildInfo'
        type: array
      name:
        type: string
    type: object
  editorapi.modelEditPara:
    properties:
      model_info:
        $ref: '#/definitions/editorapi.modelEdit'
        type: object
      model_key:
        type: string
    type: object
...
  editorapi.soundActor:
    properties:
      city:
        description: 城市
        type: string
      desc:
        description: 描述
        type: string
      gender:
        type: integer
      name:
        type: string
      photo:
        type: string
      role:
        type: string
      sound:
        type: string
    type: object
  editorapi.subFeildInfo:
    properties:
      data_from:
        type: string
      desc:
        type: string
      feild:
        type: string
      list_order:
        type: integer
      name:
        type: string
      sub_feilds:
        items:
          $ref: '#/definitions/editorapi.subFeildInfoSon'
        type: array
      type:
        type: string
    type: object
  editorapi.subFeildInfoSon:
    properties:
      data_from:
        type: string
      desc:
        type: string
      feild:
        type: string
      list_order:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
...
  requests.QRcodeImageRequests:
    properties:
      size:
        description: 二维码图片大小
        type: integer
      url:
        description: Url内容
        type: string
    required:
    - url
    type: object
...
info:
  contact: {}
  description: 全球说编辑器后台API服务文档
  license: {}
  title: 全球说编辑器后台API服务文档
  version: 0.0.1
paths:
...
  /dis/channel/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 频道数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editor.DiscoverChannel'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑频道
      tags:
      - DisChannelAPI（发现频道接口）
...
  /editor/catalog/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 修改目录
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.catalogEditParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改目录
      tags:
      - EditorCatalogAPI(目录接口)
...
  /editor/class_weixin/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: 课程微信数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.classWeixinEditPara'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"修改成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑课程微信
      tags:
      - LiveAPI（课程微信接口）
...
  /editor/content/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 编辑文件内容，Contents是内容数组，内容的字段，根据内容模型设置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.contentEditParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑文件内容
      tags:
      - EditorContentAPI（内容接口）
...
  /editor/content/version/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程内容修改
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.contentEditParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 课程内容版本修改
      tags:
      - EditorContentVersionAPI（内容版本接口）
...
  /editor/contenttype/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加类型参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.contenType'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑内容类型
      tags:
      - EditorContentTypeAPI(课程内容类型接口)
...
  /editor/course/detail/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程详情
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.CoursePriDetail'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑小学英语课程详情
      tags:
      - EditorCourseAPI（课程接口）
  /editor/course/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 修改课程
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.courseEditParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 修改课程
      tags:
      - EditorCourseAPI（课程接口）
...
  /editor/image/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片添加
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.imageAddParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片添加
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/add/more:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片添加
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.imageAddMoreParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片添加
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片删除
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.imageDelParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片删除
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/download:
    get:
      consumes:
      - multipart/form-data
      parameters:
      - description: 图片标签
        in: formData
        name: tagKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片信息下载
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片添加
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.imageEditParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片编辑
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片搜索参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.imageSearchParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片搜索
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/tag/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片标签添加
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.imageTagParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片标签添加
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/tag/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: 图片标签添加
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.imageTagParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片标签删除
      tags:
      - EditorImageAPI （图片系统接口）
  /editor/image/tags:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 图片标签列表
      tags:
      - EditorImageAPI （图片系统接口）
...
  /editor/lang/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语种信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.langEditPara'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"编辑成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑语种信息
      tags:
      - EditorLangAPI（语言种类接口）
...
  /editor/model/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语种信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.modelEditPara'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"编辑成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑语种信息
      tags:
      - EditorModelAPI（内容模型接口）
...
  /editor/qr_code/image:
    post:
      consumes:
      - application/json
      description: 生成二维码
      parameters:
      - description: 列表参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.QRcodeImageRequests'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
        "400":
          description: '{"code":500,"data":{},"msg":"获取失败信息"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 生成二维码
      tags:
      - QRcodeAPI（二维码管理）
...
  /editor/sword/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 关键词
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.SearchWord'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑关键词
      tags:
      - EditorSearchWordAPI（搜索关键词接口）
...
  /editor/tag/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editor.Content_tags'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 添加标签
      tags:
      - EditorTagAPI（标签接口）
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"